requirements.txt
astroid==2.0.4
autopep8==1.4.2
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.3
grpcio==1.16.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
logzero==1.5.0
mccabe==0.6.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pycodestyle==2.4.0
pylint==2.1.1
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.0
rope==0.11.0
rsa==4.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
webrtcvad==2.0.10
wrapt==1.10.11
